Exercises 1 and 2 - Code from Blackjack cards
This is the first exercise in chapter one of the book, where the reader must organize the code, compile it, and test it on the fly. The code is a card identifier, from Ace to King, for the game Blackjack.

The second exercise refers to the previous card counter, with the addition of a count for the Blackjack game. The code checks if the number entered by the user is between 3 and 6. If it is, it will increase the count by 1. If it is a 10, jack, queen or king, it will decrease the count by 1.